fix: IntegrationDeletedProcessor
